backend/requirements.txt
fastapi==0.129.0
uvicorn==0.41.0
python-dotenv==1.2.1
google-genai==1.64.0
pandas==3.0.1
httpx==0.28.1
